Fix for avatar imposters and a little tweakage here n there

+++ S18/indra/llrender/llvertexbuffer.cpp
@@ -811,7 +811,7 @@ U8* LLVertexBuffer::mapBuffer(S32 access)
 
 			GLint buff;
 			glGetIntegerv(GL_ARRAY_BUFFER_BINDING_ARB, &buff);
-			if ((GLuint)buff != mGLBuffer)
+			if (buff != mGLBuffer)
 			{
 				llerrs << "Invalid GL vertex buffer bound: " << buff << llendl;
 			}
@@ -824,7 +824,7 @@ U8* LLVertexBuffer::mapBuffer(S32 access)
 		{
 			GLint buff;
 			glGetIntegerv(GL_ELEMENT_ARRAY_BUFFER_BINDING_ARB, &buff);
-			if ((GLuint)buff != mGLIndices)
+			if (buff != mGLIndices)
 			{
 				llerrs << "Invalid GL index buffer bound: " << buff << llendl;
 			}
